+++ Cargo.toml
@@ -10,6 +10,7 @@ members = [
     "golem-template-service",
     "golem-shard-manager",
     "golem-worker-service-base",
+    "golem-worker-service",
 ]
 
 exclude = [
