+++ golem-worker-service-base/Cargo.toml
@@ -1,19 +1,10 @@
 [package]
-name = "golem-worker-service"
+name = "golem-worker-service-base"
 version = "0.0.0"
 edition = "2021"
 license = "Apache-2.0"
 homepage = "https://golem.cloud"
 description = "Base functionalities of Golem Worker Service"
-
-
-[[bin]]
-name = "golem-worker-service"
-path = "src/main.rs"
-
-[[bin]]
-name = "golem-worker-service-yaml"
-path = "src/openapi_yaml.rs"
 
 [lib]
