+++ golem-worker-service-base/src/lib.rs
@@ -6,7 +6,6 @@ pub mod app_config;
 pub mod auth;
 pub mod evaluator;
 pub mod expr;
-pub mod grpcapi;
 mod http_request;
 pub mod metrics;
 pub mod oas_worker_bridge;
@@ -16,7 +15,7 @@ pub mod service;
 pub mod tokeniser;
 pub mod value_typed;
 pub mod worker_request;
-pub mod worker_request_to_http_response;
+pub mod worker_response;
 pub mod worker_request_to_response;
 
 pub trait UriBackConversion {
