+++ golem-worker-service/src/grpcapi/worker.rs
