+++ golem-worker-service/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod api;
+pub mod service;
+pub mod grpcapi;
+
+pub mod worker_request_to_http_response;
