+++ golem-worker-service/src/openapi_yaml.rs
@@ -1,0 +1,23 @@
+// Copyright 2024 Golem Cloud
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use golem_worker_service::api::make_open_api_service;
+use golem_worker_service::service::Services;
+
+#[tokio::main]
+async fn main() {
+    let services = Services::noop();
+    let api_service = make_open_api_service(&services);
+    println!("{}", api_service.spec_yaml())
+}
